Updated code for PR #29
- Removed typo 'auxillary'
- Added `internal/filesystem`
- Moved type definition for configuration file to `internal/types`
- Added some reference comments for GoDoc
